Extracted hand display logic to a separate definition
